+++ bindings/mnt/pyfiction/src/pyfiction/algorithms/simulation/sidb/is_operational.cpp
@@ -42,6 +42,13 @@ void is_operational_impl(nanobind::module_& m)
           py::arg("lyt"), py::arg("spec"), py::arg("params"), py::arg("input_bdl_wire"), py::arg("output_bdl_wire"),
           py::arg("canvas_lyt") = std::nullopt, DOC(fiction_is_operational_2));
 
+    m.def("is_operational",
+          py::overload_cast<const std::vector<Lyt>&, const std::vector<py_tt>&, const fiction::is_operational_params&,
+                            const std::vector<fiction::bdl_wire<Lyt>>&, const std::vector<fiction::bdl_wire<Lyt>>&,
+                            const std::optional<Lyt>&>(&fiction::is_operational<Lyt, py_tt>),
+          py::arg("input_pattern_layouts"), py::arg("spec"), py::arg("params"), py::arg("input_bdl_wire"),
+          py::arg("output_bdl_wire"), py::arg("canvas_lyt") = std::nullopt, DOC(fiction_is_operational_3));
+
     m.def("operational_input_patterns",
           py::overload_cast<const Lyt&, const std::vector<py_tt>&, const fiction::is_operational_params&>(
               &fiction::operational_input_patterns<Lyt, py_tt>),